+++ libraries/MCP3424/MCP3424.h
@@ -2,7 +2,7 @@
 //
 //    FILE: MCP3424.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.0
+// VERSION: 0.1.1
 // PURPOSE: Arduino library for 18 bit ADC I2C MCP3424 and compatibles.
 //     URL: https://github.com/RobTillaart/MCP3424
 
@@ -10,14 +10,8 @@
 #include "Arduino.h"
 #include "Wire.h"
 
-#define MCP3424_LIB_VERSION               (F("0.1.0"))
+#define MCP3424_LIB_VERSION               (F("0.1.1"))
 
-
-//  TODO
-//  - ERROR HANDLING
-//  - READ STATUS
-//  - CACHE LAST READ
-//
 
 class MCP3424
 {
@@ -50,12 +44,11 @@ public:
   uint8_t  getMode();
 
 
-private:
+protected:
   uint8_t   _maxChannels;
   uint8_t   _channel;
   uint8_t   _gain;
   uint8_t   _bits;
-  uint8_t   _mode;
   uint8_t   _config;
 
   uint8_t   _address;
@@ -63,6 +56,50 @@ private:
 
   int       writeConfig();
   int32_t   readRaw();
+};
+
+
+/////////////////////////////////////////////////////////////////////////////
+//
+//  DERIVED CLASSES
+//
+class MCP3421 : public MCP3424
+{
+public:
+  MCP3421(uint8_t address = 0x68, TwoWire *wire = &Wire);
+};
+
+class MCP3422 : public MCP3424
+{
+public:
+  MCP3422(uint8_t address = 0x68, TwoWire *wire = &Wire);
+};
+
+class MCP3423 : public MCP3424
+{
+public:
+  MCP3423(uint8_t address = 0x68, TwoWire *wire = &Wire);
+};
+
+//
+//  max 16 bit
+//
+class MCP3426 : public MCP3424
+{
+public:
+  MCP3426(uint8_t address = 0x68, TwoWire *wire = &Wire);
+};
+
+class MCP3427 : public MCP3424
+{
+public:
+  MCP3427(uint8_t address = 0x68, TwoWire *wire = &Wire);
+};
+
+class MCP3428 : public MCP3424
+{
+public:
+  MCP3428(uint8_t address = 0x68, TwoWire *wire = &Wire);
 };