+++ libraries/MCP3424/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-09-18
+- add derived classes for MCP3421/2/3/6/7/8
+- extend unit tests.
+- fixed **getMode()** to return 1 (continuous) or 0 (single shot).
+- fixed math of **floats readVolts()** et al.
+- update readme.md
+- minor edits
+
 ## [0.1.0] - 2024-09-15
 - initial version